+++ resources/views/emails/password-changed.blade.php
@@ -1,0 +1,7 @@
+<x-mail::message>
+# Password changed
+
+The password for your account was changed succesfully.
+
+If you did not change your password, please contact us as soon as possible.
+</x-mail::message>