+++ resources/views/emails/subscribed.blade.php
@@ -1,0 +1,21 @@
+<x-mail::message>
+# You upgraded to "Mixologist" plan, cheers!
+
+You now have access to all of the following features:
+
+- Create and manage up to 10 bars
+- Add unlimited cocktail recipes
+- Invite users to your bar
+- Share your recipes with bar members
+- Manage user roles
+- Enable sharing recipes via public links
+- Add more images to recipes
+- Add unlimited cocktail collections
+- No rate limits for import actions
+- Create personal access API tokens
+
+These features are automatically activated and you don't have to do anything.
+
+Enjoy,<br>
+Karlo
+</x-mail::message>